pyproject.toml
[project]
name = "confido"
version = "0.1.0"
description = ""
authors = [
    {name = "Austin Redenbaugh",email = "43683341+AustinRedenbaugh@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.82.0,<2.0.0)",
    "twilio (>=9.6.1,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
